+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Made `startTime` and `reliability` optional for hydrogeology.
 - Removed title from prompt dialog.
 - Use standard prompt dialog for deleting boreholes.
+- Updated standard alert.
 
 ### Fixed